Add the MultiPy docker images
The MultiPy image project grew out of a pull request to the fastapi uvicorn/gunicorn containers. The images are directly compatible with the FastAPI images with one major improvement- they are multiarchetecture/multiplatform so they can easily be used to create base iamges that can run on ARM (m1, gravitron) or AMD (intel, amd) systems.